requirements.txt
boto3==1.4.0
Django==1.10
djangorestframework<3.5
requests==2.10.0
zipa==0.3.0
celery==3.1.23
celery_once==0.1.4
redis==2.10.5
hashids==1.1.0
django-hashid-field==0.1.6
json-merge-patch==0.1

